fix: address localization review feedback
Agent-authored by OpenAI Codex at Z-Lemke’s request. Split specs by class, use explicit assertions, and avoid repeated locale conversion while preserving normalized filtering.
